requirements.txt
annotated-types==0.7.0
anyio==3.7.1
beautifulsoup4==4.13.4
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
dacite==1.9.2
Deprecated==1.2.18
fastapi==0.103.2
h11==0.16.0
idna==3.10
jaconv==0.4.0
markdown-it-py==3.0.0
mdurl==0.1.2
mutagen==1.47.0
platformdirs==4.3.8
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pykakasi==2.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytube==15.0.0
RapidFuzz==3.13.0
redis==6.2.0
requests==2.32.4
rich==13.9.4
six==1.17.0
sniffio==1.3.1
soundcloud-v2==1.6.0
soupsieve==2.7
spotdl==4.2.11
spotipy==2.25.1
starlette==0.27.0
syncedlyrics==1.0.1
text-unidecode==1.3
typing-inspection==0.4.1
typing_extensions==4.14.1
Unidecode==1.4.0
urllib3==2.5.0
uvicorn==0.23.2
websockets==14.2
wrapt==1.17.2
xmltodict==0.14.2
yt-dlp==2025.6.30
ytmusicapi==1.10.3
